+++ cakeshop-api/src/test/java/com/jpmorgan/ib/caonpd/cakeshop/test/ContractServiceTest.java
@@ -1,5 +1,7 @@
 package com.jpmorgan.ib.caonpd.cakeshop.test;
 
+import com.jpmorgan.ib.caonpd.cakeshop.cassandra.entity.Transaction;
+import com.jpmorgan.ib.caonpd.cakeshop.cassandra.model.Input;
 import static com.jpmorgan.ib.caonpd.cakeshop.test.Assert.*;
 import static org.testng.Assert.*;
 
@@ -7,8 +9,8 @@ import com.jpmorgan.ib.caonpd.cakeshop.db.BlockScanner;
 import com.jpmorgan.ib.caonpd.cakeshop.error.APIException;
 import com.jpmorgan.ib.caonpd.cakeshop.model.Contract;
 import com.jpmorgan.ib.caonpd.cakeshop.model.ContractABI;
-import com.jpmorgan.ib.caonpd.cakeshop.model.Transaction;
-import com.jpmorgan.ib.caonpd.cakeshop.model.Transaction.Input;
+//import com.jpmorgan.ib.caonpd.cakeshop.model.Transaction;
+//import com.jpmorgan.ib.caonpd.cakeshop.model.Transaction.Input;
 import com.jpmorgan.ib.caonpd.cakeshop.model.TransactionResult;
 import com.jpmorgan.ib.caonpd.cakeshop.service.ContractService;
 import com.jpmorgan.ib.caonpd.cakeshop.service.ContractService.CodeType;
@@ -218,7 +220,7 @@ public class ContractServiceTest extends BaseGethRpcTest {
 	    assertEquals(val2.intValue(), 200);
 
 	    // read the previous value
-	    BigInteger valPrev = (BigInteger) contractService.read(contractAddress, abi, null, "get", null, tx.getBlockNumber()-1)[0];
+	    BigInteger valPrev = (BigInteger) contractService.read(contractAddress, abi, null, "get", null, tx.getBlockNumber().longValue() -1 )[0];
 	    assertEquals(valPrev.intValue(), 100);
 	}
